+++ proto/mrpb/state_machine_log.go
@@ -1,5 +1,0 @@
-package mrpb
-
-func (rec *StateMachineLogRecord) Validate(crc uint32) bool {
-	return rec.Crc == crc
-}